Moved ContextAwareUserService to Common and injected it to AuthorizationServer
