+++ OutOfSchool/OutOfSchool.DataAccess/BusinessEntityInterceptor.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Diagnostics;
-using OutOfSchool.Common.Models;
+using OutOfSchool.Common.Services;
 using OutOfSchool.Services.Models;
 
 namespace OutOfSchool.Services;
