+++ OutOfSchool/OutOfSchool.WebApi/Startup.cs
@@ -13,7 +13,8 @@ using Microsoft.Extensions.Http.Resilience;
 using Microsoft.Extensions.Primitives;
 using OpenIddict.Abstractions;
 using OpenIddict.Validation.AspNetCore;
-using OutOfSchool.AikomApiClient.Extensions;    
+using OutOfSchool.AikomApiClient.Extensions;
+using OutOfSchool.AuthCommon.Services;
 using OutOfSchool.BackgroundJobs.Config;
 using OutOfSchool.BackgroundJobs.Extensions.Startup;
 using OutOfSchool.BusinessLogic.Config.SearchString;
@@ -33,6 +34,7 @@ using OutOfSchool.BusinessLogic.Services.Workshops;
 using OutOfSchool.Common.Communication;
 using OutOfSchool.Common.Communication.ICommunication;
 using OutOfSchool.Common.Models;
+using OutOfSchool.Common.Services;
 using OutOfSchool.EmailSender;
 using OutOfSchool.EmailSender.Services;
 using OutOfSchool.ExternalFileStore;
